Split visualize/visualize.go
